requirements.txt
appdirs==1.4.3
asgiref==3.3.4
CacheControl==0.12.6
cachetools==4.2.2
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
fastapi==0.65.2
h11==0.12.0
h3==3.7.3
html5lib==1.0.1
httptools==0.2.0
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pydantic==1.8.2
pyparsing==2.4.6
python-dotenv==0.18.0
pytoml==0.1.21
PyYAML==5.4.1
redis==3.5.3
requests==2.22.0
retrying==1.3.3
six==1.14.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.25.8
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
webencodings==0.5.1
websockets==9.1
